winimpsyms: two passes
Move to a two-pass model in which we first scan for all import
symbols, then do the real symbol dump for both the import symbols
(e.g. "__imp_X") and for the target of the import ("X"). Treat
defs as refs and show in the refs list for each symbol.